feat(server): 升级主键为雪花ID并添加数据初始化
1. 将审计模型 CreateBy/UpdateBy 字段类型从 uint 升级为 int64，对齐雪花 ID
2. 修改通用请求模型 GetById/GetRoleId 的 ID 为 int64，IdsReq 的 Ids 为 []string 并增加 Int64s 转换方法
3. 新增 Casbin、角色、用户、菜单的数据初始化器，支持自动建表和默认数据填充
4. 调整前端环境变量超级角色标识为小写 "super"
5. 添加 Trae 规则适配层文件，指向真实规则入口

+++ server/global/model.go
@@ -23,8 +23,9 @@ type OPS_MODEL struct {
 }
 
 // OPS_AUDIT_MODEL 审计基座:内嵌 OPS_BASE + CreateBy/UpdateBy,
+// CreateBy/UpdateBy 从 uint 升级为 int64(对齐雪花 ID 主键类型,json string 传输对齐前端)
 type OPS_AUDIT_MODEL struct {
 	OPS_BASE
-	CreateBy uint `json:"createBy,string" gorm:"comment:创建者"`
-	UpdateBy uint `json:"updateBy,string" gorm:"comment:更新者"`
+	CreateBy int64 `json:"createBy,string" gorm:"comment:创建者"`
+	UpdateBy int64 `json:"updateBy,string" gorm:"comment:更新者"`
 }
